More general fixes for stability
